Refactor of returns controller plus specs
