+++ app/uk/gov/hmrc/plasticpackagingtaxreturns/controllers/ReturnsController.scala
@@ -19,28 +19,25 @@ package uk.gov.hmrc.plasticpackagingtaxreturns.controllers
 import com.google.inject.{Inject, Singleton}
 import play.api.Logger
 import play.api.libs.json.JsObject
-import play.api.libs.json.Json.{reads, toJson}
-import play.api.mvc.{Action, AnyContent, ControllerComponents}
+import play.api.libs.json.Json.toJson
 import play.api.mvc._
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.plasticpackagingtaxreturns.audit.Auditor
 import uk.gov.hmrc.plasticpackagingtaxreturns.config.AppConfig
 import uk.gov.hmrc.plasticpackagingtaxreturns.connectors.ReturnsConnector
-import uk.gov.hmrc.plasticpackagingtaxreturns.connectors.models.eis.returns.{Calculations, NrsReturnOrAmendSubmission, Return, ReturnWithNrsFailureResponse, ReturnWithNrsSuccessResponse, ReturnsSubmissionRequest}
+import uk.gov.hmrc.plasticpackagingtaxreturns.connectors.models.eis.returns._
 import uk.gov.hmrc.plasticpackagingtaxreturns.controllers.actions.{Authenticator, AuthorizedRequest}
 import uk.gov.hmrc.plasticpackagingtaxreturns.controllers.response.JSONResponses
 import uk.gov.hmrc.plasticpackagingtaxreturns.models.ReturnType.ReturnType
-import uk.gov.hmrc.plasticpackagingtaxreturns.models.{ReturnType, ReturnValues, TaxReturn}
-import uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.UserAnswers
 import uk.gov.hmrc.plasticpackagingtaxreturns.models.nonRepudiation.{NonRepudiationSubmissionAccepted, NrsDetails}
+import uk.gov.hmrc.plasticpackagingtaxreturns.models.{ReturnType, ReturnValues}
 import uk.gov.hmrc.plasticpackagingtaxreturns.repositories.SessionRepository
 import uk.gov.hmrc.plasticpackagingtaxreturns.services.PPTReturnsCalculatorService
 import uk.gov.hmrc.plasticpackagingtaxreturns.services.nonRepudiation.NonRepudiationService
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 
 import java.time.format.DateTimeFormatter
-import java.time.{LocalDate, LocalDateTime, ZoneId, ZoneOffset, ZonedDateTime}
-import java.util.Locale
+import java.time.{ZoneId, ZonedDateTime}
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
 
@@ -87,20 +84,24 @@ class ReturnsController @Inject()(
           .fold {
             Future.successful(UnprocessableEntity("Unable to build ReturnsSubmissionRequest from UserAnswers"))
           } { case (returnValues, userAnswers) =>
-          val calculations: Calculations = calculationsService.calculate(returnValues)
-          val eisRequest: ReturnsSubmissionRequest = ReturnsSubmissionRequest(returnValues, calculations, submissionId = submissionId, returnType = returnType)
+            val calculations: Calculations = calculationsService.calculate(returnValues)
 
-          returnsConnector.submitReturn(pptReference, eisRequest).flatMap {
-            case Right(response) =>
-              sessionRepository.clear(request.cacheKey).andThen {
-                case Success(_)  => logger.info(s"Successfully removed tax return for $pptReference from cache")
-                case Failure(ex) => logger.warn(s"Failed to remove tax return for $pptReference from cache- ${ex.getMessage}", ex)
+            if (calculations.isSubmittable) {
+              val eisRequest: ReturnsSubmissionRequest = ReturnsSubmissionRequest(returnValues, calculations, submissionId, returnType)
+
+              returnsConnector.submitReturn(pptReference, eisRequest).flatMap {
+                case Right(response) =>
+                  sessionRepository.clear(request.cacheKey).andThen {
+                    case Success(_) => logger.info(s"Successfully removed tax return for $pptReference from cache")
+                    case Failure(ex) => logger.warn(s"Failed to remove tax return for $pptReference from cache- ${ex.getMessage}", ex)
+                  }
+                  handleNrsRequest(request, userAnswers.data, eisRequest, response)
+                case Left(errorStatusCode) => Future.successful(new Status(errorStatusCode))
               }
-              handleNrsRequest(request, userAnswers.data, eisRequest, response)
-            case Left(errorStatusCode) => Future.successful(new Status(errorStatusCode))
+            } else {
+              Future.successful(UnprocessableEntity("The calculation is not submittable"))
+            }
           }
-
-        }
       }
 
     }
