+++ test/uk/gov/hmrc/plasticpackagingtaxreturns/controllers/builders/ReturnsSubmissionResponseBuilder.scala
@@ -18,7 +18,6 @@ package uk.gov.hmrc.plasticpackagingtaxreturns.controllers.builders
 
 import uk.gov.hmrc.plasticpackagingtaxreturns.connectors.models.eis.returns.{EisReturnDetails, IdDetails, Return, ReturnWithNrsFailureResponse, ReturnWithNrsSuccessResponse}
 
-import java.time.LocalDate
 
 trait ReturnsSubmissionResponseBuilder {
